+++ @narative/gatsby-theme-novela/gatsby-config.js
@@ -54,7 +54,7 @@ module.exports = ({
         },
         feeds: [
           {
-            serialize: ({ query: { site, allArticle, allContentfulPost } }) => {
+            serialize: ({ query: { site, allArticle, allContentfulArticle } }) => {
               if (local && !contentful) {
                 return allArticle.edges
                   .filter(edge => !edge.node.secret)
@@ -65,26 +65,27 @@ module.exports = ({
                       date: edge.node.date,
                       url: site.siteMetadata.siteUrl + edge.node.slug,
                       guid: site.siteMetadata.siteUrl + edge.node.slug,
+                      // body is raw JS and MDX; will need to be processed before it can be used
                       // custom_elements: [{ "content:encoded": edge.node.body }],
                       author: edge.node.author,
                     };
                   });
               } else if (!local && contentful) {
-                return allContentfulPost.edges
+                return allContentfulArticle.edges
                   .filter(edge => !edge.node.secret)
                   .map(edge => {
                     return {
                       ...edge.node,
                       description: edge.node.excerpt,
                       date: edge.node.date,
-                      url: site.siteMetadata.siteUrl + edge.node.slug,
-                      guid: site.siteMetadata.siteUrl + edge.node.slug,
-                      // custom_elements: [{ "content:encoded": edge.node.body }],
-                      author: edge.node.author,
+                      url: site.siteMetadata.siteUrl + '/' + edge.node.slug,
+                      guid: site.siteMetadata.siteUrl + '/' + edge.node.slug,
+                      custom_elements: [{ "content:encoded": edge.node.body.childMarkdownRemark.html }],
+                      author: edge.node.author ? edge.node.author.name : '',
                     };
                   });
               } else {
-                const allArticlesData = { ...allArticle, ...allContentfulPost };
+                const allArticlesData = { ...allArticle, ...allContentfulArticle };
                 return allArticlesData.edges
                   .filter(edge => !edge.node.secret)
                   .map(edge => {
@@ -95,7 +96,7 @@ module.exports = ({
                       url: site.siteMetadata.siteUrl + edge.node.slug,
                       guid: site.siteMetadata.siteUrl + edge.node.slug,
                       // custom_elements: [{ "content:encoded": edge.node.body }],
-                      author: edge.node.author,
+                      author: edge.node.author ? edge.node.author.name : '',
                     };
                   });
               }
@@ -107,6 +108,7 @@ module.exports = ({
                 allArticle(sort: {order: DESC, fields: date}) {
                   edges {
                     node {
+                      body
                       excerpt
                       date
                       slug
@@ -121,13 +123,18 @@ module.exports = ({
                 : !local && contentful
                 ? `
               {
-                allContentfulPost(sort: {order: DESC, fields: date}) {
+                allContentfulArticle(sort: {order: DESC, fields: date}) {
                   edges {
                     node {
                       excerpt
                       date
                       slug
                       title
+                      body {
+                        childMarkdownRemark {
+                          html
+                        }
+                      }
                       author {
                         name
                       }
@@ -142,6 +149,7 @@ module.exports = ({
                 allArticle(sort: {order: DESC, fields: date}) {
                   edges {
                     node {
+                      body
                       excerpt
                       date
                       slug
@@ -151,13 +159,18 @@ module.exports = ({
                     }
                   }
                 }
-                allContentfulPost(sort: {order: DESC, fields: date}) {
+                allContentfulArticle(sort: {order: DESC, fields: date}) {
                   edges {
                     node {
                       excerpt
                       date
                       slug
                       title
+                      body {
+                        childMarkdownRemark {
+                          html
+                        }
+                      }
                       author {
                         name
                       }