+++ @narative/gatsby-theme-novela/src/gatsby/node/createPages.js
@@ -60,6 +60,16 @@ module.exports = async ({ actions: { createPage }, graphql }, themeOptions) => {
     tags = false,
   } = themeOptions;
 
+  const { data } = await graphql(`
+    query siteQuery {
+      site {
+        siteMetadata {
+          siteUrl
+        }
+      }
+    }
+  `);
+
   console.log(sources);
   // Defaulting to look at the local MDX files as sources.
   const { local = true, contentful = false } = sources;
@@ -215,6 +225,7 @@ module.exports = async ({ actions: { createPage }, graphql }, themeOptions) => {
         article,
         authors: authorsThatWroteTheArticle,
         basePath,
+        permalink: `${data.site.siteMetadata.siteUrl}${article.slug}/`,
         slug: article.slug,
         id: article.id,
         title: article.title,