+++ @narative/gatsby-theme-novela/src/gatsby/node/onCreateNode.js
@@ -48,7 +48,11 @@ module.exports = ({ node, actions, getNode, createNodeId }, themeOptions) => {
   // ///////////////////////////////////////////////////////
 
   if (node.internal.type === `AuthorsYaml`) {
-    const slug = node.slug ? `/${node.slug}` : slugify(node.name);
+    const slug = node.slug
+      ? `/${node.slug}`
+      : slugify(node.name, {
+          lower: true,
+        });
 
     const fieldData = {
       ...node,
@@ -87,7 +91,9 @@ module.exports = ({ node, actions, getNode, createNodeId }, themeOptions) => {
       slug: generateSlug(
         basePath,
         generateArticlePermalink(
-          slugify(node.frontmatter.slug || node.frontmatter.title),
+          slugify(node.frontmatter.slug || node.frontmatter.title, {
+            lower: true,
+          }),
           node.frontmatter.date,
         ),
       ),
@@ -121,7 +127,13 @@ module.exports = ({ node, actions, getNode, createNodeId }, themeOptions) => {
     createNodeField({
       node,
       name: `slug`,
-      value: generateSlug(basePath, 'authors', slugify(node.name)),
+      value: generateSlug(
+        basePath,
+        'authors',
+        slugify(node.name, {
+          lower: true,
+        }),
+      ),
     });
 
     createNodeField({