+++ www/gatsby-config.js
@@ -38,6 +38,16 @@ const siteMetadata = {
     {
       url: `https://digitalocean.com`,
     },
+    {
+      url: `https://tripadvisor.com`,
+    },
+    {
+      url: `https://notion.so`,
+    },
+    {
+      name: `mailto`,
+      url: `mailto:`,
+    },
   ],
 };